deduplicating code, marking for future plugin api refactorings.
