+++ src/modules/shared/README.md
@@ -1,0 +1,3 @@
+## modules/shared
+
+These are common modules/reducers that should move to a common plugin module.
