+++ src/modules/shared/global-app-registry.js
