+++ src/modules/shared/index.js
@@ -1,0 +1,17 @@
+import ns, { nsChanged } from './ns';
+import dataService, { dataServiceConnected } from './data-service';
+import appRegistry, { appRegistryActivated } from './app-registry';
+import globalAppRegistry, {
+  globalAppRegistryActivated
+} from './global-app-registry';
+
+export {
+  ns,
+  nsChanged,
+  dataService,
+  dataServiceConnected,
+  appRegistry,
+  appRegistryActivated,
+  globalAppRegistry,
+  globalAppRegistryActivated
+};
